+++ lib/active_model_serializers/deserialization.rb
@@ -3,11 +3,11 @@ module ActiveModelSerializers
     module_function
 
     def jsonapi_parse(*args)
-      ActiveModelSerializers::Adapter::JsonApi::Deserialization.parse(*args)
+      Adapter::JsonApi::Deserialization.parse(*args)
     end
 
     def jsonapi_parse!(*args)
-      ActiveModelSerializers::Adapter::JsonApi::Deserialization.parse!(*args)
+      Adapter::JsonApi::Deserialization.parse!(*args)
     end
   end
 end
